[blockchain] change type in nm for dfsstate

+++ headers/managers/data_mining_manager.h
@@ -35,8 +35,8 @@ public:
                         const bool isHahsing, std::string &result);
     std::string rootMerkleHash(std::string &data);
 
-    double calculateCoins(int dataAmountStored, int dataAmountTotalStoredInNetwork, int circulativeSupply,
-                          int blockAmount, double coefficient);
+    double calculateCoins(uint64_t dataAmountStored, uint64_t dataAmountTotalStoredInNetwork, uint64_t circulativeSupply,
+                          uint64_t blockAmount, double coefficient);
 private:
     std::vector<MerkleDataBlocks> splitListIntoPair(std::vector<std::string> &vector, const bool isHahsing);
     void hashingElements(std::vector<std::string> &vector);
